Career questsession one brians notes

+++ FlightBook/app/forms/create_site.py
@@ -1,13 +1,13 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, IntegerField, DateTimeField, TextAreaField, BooleanField
+from wtforms import StringField, IntegerField, DateTimeField, TextAreaField, BooleanField, FloatField
 from wtforms.validators import DataRequired, Optional, Length, NumberRange
 from flask_wtf.file import FileField, FileAllowed
 from ..api.aws_functions import ALLOWED_EXTENSIONS
 
 class SiteCreateForm(FlaskForm):
     name = StringField('Site Name', validators=[DataRequired(), Length(min=7,max=40)])
-    lat = IntegerField('latitude')
-    lon = IntegerField('longitude')
+    lat = FloatField('latitude')
+    lon = FloatField('longitude')
     altitude = IntegerField('altitude', validators=[NumberRange(min=0, max=17999)])
     intro = TextAreaField('Log', validators=[DataRequired(), Length(max=1024)])
     user_id = IntegerField('user_id')
